feat: resolve issues #121, #122, #194, #195
- #121: Makefile already present with all required targets
- #122: deploy.sh already had strict error handling, colors, --network flag, .env.local write, and summary table
- #194: split contracts/crowdfund/src/lib.rs into types.rs, errors.rs, storage.rs modules
- #195: fix eslint.config.js to ESLint 9 flat config with jsx-a11y/recommended; fix label-has-associated-control in dashboard/page.tsx; fix broken JSX in PledgeModal.tsx and suppress backdrop click-without-keyboard with explanation comment
