+++ contracts/crowdfund/src/lib.rs
@@ -2,165 +2,15 @@
 #![allow(missing_docs)]
 #![allow(clippy::too_many_arguments)]
 
-use soroban_sdk::{
-    contract, contractimpl, contracttype, symbol_short, token, Address, Env, String, Symbol, Vec,
-};
+mod errors;
+mod storage;
+mod types;
 
-const CONTRACT_VERSION: u32 = 3;
+pub use errors::ContractError;
+pub use storage::{CONTRACT_VERSION, KEY_ADMIN, KEY_CONTRIBS, KEY_CREATOR, KEY_DEADLINE, KEY_DESC, KEY_GOAL, KEY_MIN, KEY_PLATFORM, KEY_SOCIAL, KEY_STATUS, KEY_TITLE, KEY_TOKEN, KEY_TOTAL};
+pub use types::{CampaignInfo, CampaignStats, DataKey, PlatformConfig, Status};
 
-// ── Storage Keys ──────────────────────────────────────────────────────────────
-const KEY_CREATOR: Symbol = symbol_short!("CREATOR");
-const KEY_TOKEN: Symbol = symbol_short!("TOKEN");
-const KEY_GOAL: Symbol = symbol_short!("GOAL");
-const KEY_DEADLINE: Symbol = symbol_short!("DEADLINE");
-const KEY_TOTAL: Symbol = symbol_short!("TOTAL");
-const KEY_CONTRIBS: Symbol = symbol_short!("CONTRIBS");
-const KEY_STATUS: Symbol = symbol_short!("STATUS");
-const KEY_MIN: Symbol = symbol_short!("MIN");
-const KEY_TITLE: Symbol = symbol_short!("TITLE");
-const KEY_DESC: Symbol = symbol_short!("DESC");
-const KEY_SOCIAL: Symbol = symbol_short!("SOCIAL");
-const KEY_PLATFORM: Symbol = symbol_short!("PLATFORM");
-const KEY_ADMIN: Symbol = symbol_short!("ADMIN");
-
-// ── Data Types ────────────────────────────────────────────────────────────────
-
-/// Campaign status enumeration.
-///
-/// Represents the lifecycle state of a crowdfunding campaign.
-#[derive(Clone, PartialEq, Debug)]
-#[contracttype]
-pub enum Status {
-    /// Campaign is accepting contributions
-    Active,
-    /// Campaign deadline passed and goal was reached
-    Successful,
-    /// Campaign deadline passed and goal was not reached (refunds available)
-    Refunded,
-    /// Campaign was cancelled by creator (refunds available)
-    Cancelled,
-    /// Campaign is temporarily paused (no new contributions allowed)
-    Paused,
-}
-
-/// Campaign statistics snapshot.
-///
-/// Contains aggregated metrics about campaign progress and contributor activity.
-#[derive(Clone)]
-#[contracttype]
-pub struct CampaignStats {
-    /// Total amount raised in stroops
-    pub total_raised: i128,
-    /// Campaign funding goal in stroops
-    pub goal: i128,
-    /// Progress as basis points (0-10000, where 10000 = 100%)
-    pub progress_bps: u32,
-    /// Number of unique contributors
-    pub contributor_count: u32,
-    /// Average contribution amount in stroops (total_raised / contributor_count)
-    pub average_contribution: i128,
-    /// Largest single contribution amount in stroops
-    pub largest_contribution: i128,
-}
-
-/// Platform fee configuration.
-///
-/// Specifies the address that receives platform fees and the fee percentage.
-#[derive(Clone, PartialEq, Debug)]
-#[contracttype]
-pub struct PlatformConfig {
-    /// Address that receives platform fees
-    pub address: Address,
-    /// Fee percentage in basis points (e.g., 250 = 2.5%)
-    pub fee_bps: u32,
-}
-
-/// Complete campaign information.
-///
-/// Contains all metadata and configuration for a campaign.
-#[derive(Clone)]
-#[contracttype]
-pub struct CampaignInfo {
-    /// Campaign creator's Stellar address
-    pub creator: Address,
-    /// Token address for contributions
-    pub token: Address,
-    /// Funding goal in stroops
-    pub goal: i128,
-    /// Campaign deadline as Unix timestamp (seconds)
-    pub deadline: u64,
-    /// Minimum contribution amount in stroops
-    pub min_contribution: i128,
-    /// Campaign title
-    pub title: String,
-    /// Campaign description
-    pub description: String,
-    /// Current campaign status
-    pub status: Status,
-    /// Whether a platform fee is configured
-    pub has_platform_config: bool,
-    /// Platform fee in basis points (0 if no config)
-    pub platform_fee_bps: u32,
-    /// Platform fee recipient address
-    pub platform_address: Address,
-}
-
-/// Storage key variants for contract data.
-///
-/// Used to organize persistent and instance storage in the contract.
-#[derive(Clone)]
-#[contracttype]
-pub enum DataKey {
-    /// Contribution amount for a specific address
-    Contribution(Address),
-    /// Whether an address has contributed (presence flag)
-    ContributorPresence(Address),
-    /// Total number of unique contributors
-    ContributorCount,
-    /// Largest single contribution amount
-    LargestContribution,
-    /// Whitelist of accepted token addresses
-    AcceptedTokens,
-}
-
-// ── Contract Errors ───────────────────────────────────────────────────────────
-
-use soroban_sdk::contracterror;
-
-/// Contract error types.
-///
-/// Represents all possible error conditions that can occur during contract execution.
-#[contracterror]
-#[derive(Copy, Clone, Debug, Eq, PartialEq, PartialOrd, Ord)]
-#[repr(u32)]
-pub enum ContractError {
-    /// Campaign has already been initialized
-    AlreadyInitialized = 1,
-    /// Campaign deadline has passed
-    CampaignEnded = 2,
-    /// Campaign deadline has not yet passed
-    CampaignStillActive = 3,
-    /// Campaign goal was not reached
-    GoalNotReached = 4,
-    /// Campaign goal was already reached
-    GoalReached = 5,
-    /// Arithmetic overflow occurred
-    Overflow = 6,
-    /// Campaign is not in Active status
-    NotActive = 7,
-    /// Platform fee is invalid (> 10,000 bps)
-    InvalidFee = 8,
-    /// Amount is below minimum contribution
-    BelowMinimum = 9,
-    /// Deadline is invalid
-    InvalidDeadline = 10,
-    /// Campaign is paused
-    CampaignPaused = 11,
-    /// Campaign goal is invalid (<= 0)
-    InvalidGoal = 12,
-    /// Token is not accepted by this campaign
-    TokenNotAccepted = 13,
-}
+use soroban_sdk::{contract, contractimpl, token, Address, Env, String, Vec};
 
 // ── Contract ──────────────────────────────────────────────────────────────────
